rewo

+++ src/main/java/me/jweissen/aeticket/dto/request/CategoryRequestDto.java
@@ -1,4 +1,4 @@
 package me.jweissen.aeticket.dto.request;
 
-public record CategoryRequestDto(String name, int price, int stock) {
+public record CategoryRequestDto(String name, Double price, Integer stock) {
 }
